requirements.txt
fastapi==0.109.0
python-multipart==0.0.6
Pillow==10.2.0
pytesseract==0.3.10
google-cloud-vision==3.4.5
uvicorn==0.25.0
